media-source/SourceBufferList-length.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Check the number of created SourceBuffer objects in the list</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var FILE = 'test.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';

//check the browser supports the MIME used in this test
function canPlayType(mime) {
    if(video.canPlayType(mime) != 'probably') {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test.");
        });
        this.done();
        return false;
    }
    return true;
}
async_test(function(t) {
    if(!canPlayType.bind(t)(mime)) {
        return;
    }
    var mediaSource = new MediaSource();
    mediaSource.addEventListener('sourceopen', function(e){
        t.step(function() {
            assert_equals(mediaSource.sourceBuffers.length, 0);
            var sourceBuffer = mediaSource.addSourceBuffer(mime);
            assert_equals(mediaSource.sourceBuffers.length, 1);
        });
        t.done();
    });
    video.src = window.URL.createObjectURL(mediaSource);
},
'SourceBufferList#length : Check the number of created SourceBuffer objects in the list.');
</script>
</body>
</html>
